requirements.txt
numpy
scipy
nibabel
torch
torchvision
torchaudio
tqdm
scikit-learn
matplotlib
seaborn
pyyaml
rich
